Filter out updates that come from the future.
If clock skew means that updates are reported as coming from the future,
(more than the update interval), then those updates will be reported
multiple times in the notification bubbles.

To avoid this we filter out updates that come from the future, and
wait until our clock has caught up to include them. It will mean a
delay in some updates being shown, but it avoid having to maintain
a cache of notified updates, which would be a pain.